Separating email-validator code from email-provider-list code.
