+++ src/lib/io/live-downloader.ts
@@ -2,7 +2,8 @@ import DataDir from '../cache/datadir.js';
 import config from '../config/index.js';
 import { EntityKind, FullEntity } from '../model/hubspot/interfaces.js';
 import { RawLicense, RawTransaction } from "../model/marketplace/raw";
-import { downloadAllTlds, downloadFreeEmailProviders } from '../services/domains.js';
+import { downloadAllTlds } from '../services/domains.js';
+import { downloadFreeEmailProviders } from '../services/email-providers.js';
 import LiveHubspot from '../services/hubspot.js';
 import { Marketplace } from '../services/marketplace.js';
 import { Downloader, Progress } from './interfaces.js';
